Feature#5

See merge request edu/aizhixin_api_cloud!136

+++ diandian/src/main/java/com/aizhixin/cloud/dd/feedback/domain/FeedbackRecordDomain.java
@@ -52,10 +52,10 @@ public class FeedbackRecordDomain {
     private String userAvatar;
 
     @ApiModelProperty(value = "教师评价总分")
-    private Integer teachingScore;
+    private Float teachingScore;
 
     @ApiModelProperty(value = "学风评价总分")
-    private Integer studyStyleScore;
+    private Float studyStyleScore;
 
     @ApiModelProperty(value = "创建时间")
     private Date createDate;
@@ -76,7 +76,7 @@ public class FeedbackRecordDomain {
 
     }
 
-    public FeedbackRecordDomain(Long id, Long templetId, String teachingClassId, String teachingClassCode, String teachingClassName, String teachingClassTeacher, String teacherJobNum, String courseId, String courseName, String jobNum, String userName, String userAvatar, Integer teachingScore, Integer studyStyleScore) {
+    public FeedbackRecordDomain(Long id, Long templetId, String teachingClassId, String teachingClassCode, String teachingClassName, String teachingClassTeacher, String teacherJobNum, String courseId, String courseName, String jobNum, String userName, String userAvatar, Float teachingScore, Float studyStyleScore) {
         this.id = id;
         this.templetId = templetId;
         this.teachingClassId = teachingClassId;
