+++ diandian/src/main/java/com/aizhixin/cloud/dd/feedback/dto/FeedbackTempletDTO.java
@@ -26,7 +26,7 @@ public class FeedbackTempletDTO {
     private Integer quesType;
 
     @ApiModelProperty(value = "打分题总分")
-    private Integer totalScore;
+    private Float totalScore;
 
     @ApiModelProperty(value = "教师评价列表")
     private List<FeedbackTempletQuesDTO> teacherQuesList;
@@ -41,7 +41,7 @@ public class FeedbackTempletDTO {
 
     }
 
-    public FeedbackTempletDTO(Long id, Long orgId, Integer quesType, Integer totalScore) {
+    public FeedbackTempletDTO(Long id, Long orgId, Integer quesType, Float totalScore) {
         this.id = id;
         this.orgId = orgId;
         this.quesType = quesType;
